WIP: Clone repos as needed, drop "common"role
The current Ansible code assumes that development environments will only
be created in Vagrant VMs, where `~/devel` is present and contains
several git repositories. This works when one is using the vagrant file
in the root of this repository, but that assumption prevents people from
deploying development environments on non-Vagrant VMs. That's unhelpful.

In cases where a git repo must be present, use the git module to ensure
that that repo is present. This should have no effect on the use case of
creating a development environment on a Vagrant VM, and it should enable
the use case of creating a development environment on a non-Vagrant VM.

TODO: Ensure that the git module doesn't mess with existing repos, by
doing things such as changing branches. If a git repository is already
present, don't touch it.