Change how to perform actions as pulp user
Currently, when an Ansible task is to perform some task as the pulp
user, that task is marked with `become: false`. This is fragile. This
logic works if the user one is logging in as is the pulp user, but it
breaks if one is logging in as a non-pulp user, such as root.

Use `become_user: '{{ pulp_user }}'` instead. This is much more robust.
If Ansible is already connected as `pulp_user`, then Ansible won't try
to log in as `pulp_user` again, and it performs the relevant task as
`pulp_user`. And if Ansible is connected as some other user (like root),
then Ansible will become the correct user.